pyproject.toml
[project]
name = "video-ex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.116.1",
	"chardet>=5.0.0,<6",
	"langchain-core>=0.3.75",
	"langchain-google-genai>=2.1.12",
	"openpyxl>=3.1.5",
	"opencv-python>=4.12.0.88",
	"paddleocr>=3.2.0",
	"pandas>=2.3.2",
	"pillow>=11.3.0",
	"pydantic>=2.11.9",
	"python-dotenv>=1.1.1",
	"qdrant-client>=1.15.1",
	"requests>=2.32.5",
	"sarvamai>=0.1.16",
	"sentence-transformers>=5.1.1",
	"openai-whisper>=20250625",
	"uvicorn[standard]>=0.35.0",
	"psycopg2-binary>=2.9.11",
	"sqlalchemy>=2.0.48",
	"unstructured[all-docs]>=0.22.6",
	"pytesseract>=0.3.13",
]

